+++ DataClass/Quest/Quest_Complete.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SpartaTextRPG.DataClass.Quest
+{
+    //완료한 퀘스트
+    public partial class QuestList
+    {
+    }
+}
