+++ Guild.cs
@@ -1,4 +1,5 @@
 ﻿using SpartaTextRPG.DataClass;
+using SpartaTextRPG.DataClass.Quest;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -40,6 +41,7 @@ namespace SpartaTextRPG
 
                 //퀘스트 확인
                 case "2":
+                    ShowCurQuestPage();
                     break;
 
                 //퀘스트 보상
@@ -48,19 +50,37 @@ namespace SpartaTextRPG
             }
         }
 
-        //의뢰 페이지
+        //퀘스트 받기
         public void ShowRequestPage()
         {
             Console.Clear();
             Color.ChangeTextColor(Colors.YELLOW, "", "길드 - 퀘스트 받기\n");
             Console.WriteLine("새로운 퀘스트를 받을 수 있습니다.\n");
 
-            //메인 퀘스트 목록 보여주기
-            QuestList.instance.ShowQuestList();
+            //시작 가능한 퀘스트 목록 보여주기
+            QuestList.instance.ShowStQuestList();
+
+            Color.ChangeTextColor(Colors.YELLOW,"", "\n0. ");
+            Console.WriteLine("나가기");
 
             Console.WriteLine("\n의뢰 받을 퀘스트의 번호를 입력해주세요.");
             Title.PrintInputCursor();
+            Console.ReadLine();
+        }
 
+        //진행중인 퀘스트
+        public void ShowCurQuestPage()
+        {
+            Console.Clear();
+            Color.ChangeTextColor(Colors.YELLOW, "", "길드 - 퀘스트 진행\n");
+            Console.WriteLine("진행중인 퀘스트를 확인할 수 있습니다.\n");
+
+            //진행중인 퀘스트 목록 보여주기
+            QuestList.instance.ShowCurQuestList();
+            
+
+            Console.WriteLine("\n원하시는 행동을 입력해주세요.");
+            Title.PrintInputCursor();
             Console.ReadLine();
         }
     }
